Create photo puzzle in java
The program will open a window titled "Photo Puzzle."

It will display the image you specified as "your_image.jpg" divided into a grid of puzzle pieces.

The puzzle pieces will be shuffled at regular intervals (every 1000 milliseconds) due to the timer.

Users can interact with the puzzle by manually moving the puzzle pieces to solve it.

The program will continuously shuffle the puzzle pieces at the specified interval.

When the user closes the window or exits the program, it will terminate.

Please replace "your_image.jpg" with the actual path to the image you want to use for the puzzle. You can interact with the puzzle by clicking and dragging the puzzle pieces within the window to complete the image.